requirements.txt
pytorch=2.3.1
transformers=4.48.0
peft==0.14.0
numpy==1.26.4
